examples/langchain-anthropic/requirements.txt
langchain
langchain-anthropic
python-dotenv
# install humanlayer
humanlayer==0.5.6
# or if you want an editable version
# -e ../../
